:bug: [RAG] Fix error wtih ipv6 lookup
This changes fixes a bug that was caused by the IPv6
database record strings not being converted to unicode
prior to being converted to IPv6 objects.

This change also introduces tests to validate the bug fix.

This change was randomly selected as a part of the Random
Acts of Git [RAG] project by @tatemz | https://twitch.tv/tatemz