:whale: [RAG] Add Docker support for Python 2 & 3
This change introduces a Docker image configuration
for running and testing this project inside of a
containerized environment. This enables developers
to choose the version of Python they wish to test
and develop against (independent of their workstation's
version of python.

Additionally, this change adds the appropriate python
modules that are required for Python version 2. This
change modified the requirements.txt file to be a symlink
that points to the latest of two new files for Python 2
and 3 respectively.

Documentation for working with the Dockerized environment
is also added within this change and can be found in the
README files.

This change was randomly selected as a part of the Random
Acts of Git [RAG] project by @tatemz | https://twitch.tv/tatemz