:bug: [RAG] Fix support for both Python 2 & 3
This change implements a check for Python 2
in order to convert the ip address record to
unicode only within the Python 2 environment.
Otherwise, a plain string should be used within
Python 3.

This change was randomly selected as a part of the Random
Acts of Git [RAG] project by @tatemz | https://twitch.tv/tatemz