:wrench: :recycle: [RAG] Restructure project configuration
This change re-implements a single requirements.txt file in
favor of differentiating the various dependencies across
Python 2 and 3 within the setup.py configuration file.

Additionally, this change adds configuration for using the
setup.py script as a test runner that also reports coverage.

This change was randomly selected as a part of the Random
Acts of Git [RAG] project by @tatemz | https://twitch.tv/tatemz